feat(mb): separated arm lookup and point in polygon functions (Ref #96)
